+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased][]
+- Replace deprecated `componentWillMount()` with `componentDidMount()`
 
 ## [1.0.8][] - 2019-04-08
 ### Fixed